added the beneficiary grid, bind it with redux, populate values from api, fixed all the comments on draft PR.

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Request/BeneficiaryInquiry/BeneficiaryRequestDto.cs
@@ -9,5 +9,5 @@ namespace Demoulas.ProfitSharing.Common.Contracts.Request.BeneficiaryInquiry;
  public record BeneficiaryRequestDto: SortedPaginationRequestDto
 {
     public int BadgeNumber { get; set; }
-    public int PsnSuffix { get; set; }
+    public short PsnSuffix { get; set; }
 }
